pyproject.toml
[build-system]
requires = ["hatchling>=1.22.0"]
build-backend = "hatchling.build"

[project]
name = "dicom-deid-gui"
version = "0.1.0"
description = "Cross-platform HIPAA-safe DICOM de-identification utility with GUI and CLI"
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [
  { name = "DCMScrubber", email = "noreply@example.com" }
]
keywords = ["DICOM", "de-identification", "HIPAA", "PySide6", "pydicom"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Environment :: Console",
  "Environment :: X11 Applications :: Qt",
  "Intended Audience :: Healthcare Industry",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Topic :: Scientific/Engineering :: Medical Science Apps.",
]
dependencies = [
  "pydicom>=2.4.4",
  "numpy>=1.24",
  "PySide6>=6.5",
  "pylibjpeg>=2.0.0",
  "pylibjpeg-libjpeg>=2.0.0",
  "Pillow>=10.0.0",
  "rich>=13.7",
  "tqdm>=4.66",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "black>=24.3",
  "ruff>=0.5.0",
  "mypy>=1.8.0",
]

[project.scripts]
dicom-deid = "dicom_deid_gui.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["dicom_deid_gui"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = []

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
strict_optional = true
warn_return_any = true
warn_unused_ignores = true
warn_redundant_casts = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true

